Rework header processing to support uint16_t IDs

+++ src/AppState.cpp
@@ -23,7 +23,7 @@ namespace kx {
 	FilterMode g_packetFilterMode = FilterMode::ShowAll;
 
 	// Initialize filter maps (typically populated by UI setup)
-	std::map<std::pair<PacketDirection, uint8_t>, bool> g_packetHeaderFilterSelection;
+	std::map<std::pair<PacketDirection, uint16_t>, bool> g_packetHeaderFilterSelection;
 	std::map<InternalPacketType, bool> g_specialPacketFilterSelection;
 
 	// Direction Filtering
